WebService Update:
User can now verify and change password with OTP verification
blood pressure can now be implemented and some minor details in other methods

+++ Hart_Check_Official/Models/BloodPressure.cs
@@ -7,8 +7,8 @@ namespace Hart_Check_Official.Models
         [Key]
         public int bloodPressureID { get; set; }
         public int patientID { get; set; }
-        public int systolic { get; set; }
-        public int diastolic { get; set; }
+        public double systolic { get; set; }
+        public double diastolic { get; set; }
         public DateTime? dateTaken { get; set; }
 
         public Patients patients { get; set; }
